gcp_functions/requirements.txt
boto3==1.21.24
google-cloud-storage==1.7.0
